Minimize game logic dependencies
Most game logic handling used to inject the context object across the
entire calls stack. This change generally refactors game logic handling
to only inject the object it into methods which need access to ctx.bot,
and to instead pass ctx.bot.game when the only manipulations needed are
to the game. It's possible the dependency tree can be simplified
significantly beyond this, but this is still a large improvement.

This may have created issues (or made existing issues more obvious) in
creating the seating message at the beginning of the game, as Game is
generally None during this process. In particular, if a character ever
has default effects that rely on Game being defined, the generation may
break.