Add Minion Info and Demon Info to the first night
The Night object now stores its order as characters, not as players,
because _MinionInfo and _DemonInfo implement can easily implement
NightOrderMember, and there's never a reason to access part of Player
other than Player.character. These objects have to be manually appended
to the night order in its setter, which is obviously a little awkward
and will make it difficult to add new such objects in the future.