Fix incorrect night reporting scope
There was previously a bug where, if the night started in a public
context, the bot would queue the first night action (including
character-related information) in that context. The safe_bug_report
check ensures this cannot happen.

This also fixes an issue with early day starts by inserting a final
night object into the night order which has no purpose other than to
allow the storytellers to delay the night arbitrarily.

Finally, _MinionInfo and _DemonInfo were moved to Night.py instead of
Script.py, because their only function in Script is to add their names
to the script order, which is easily handled manually. Correspondingly,
the NightOrderMember ABC no longer needs a name attribute.