Fix SeesTwo _Correct pickling
Previously, because this class wasn't defined globally, it couldn't be
pickled properly. This change replaces the specific name of the effect
with "Correct", which is slightly worse for the grimoire, but means the
class can be defined globally.

This is yet another reason to switch to SQL instead of pickle.