Add context type hints with non-optional game
There were a number of mypy erros being registered because ctx.bot.game
may be None (and ditto with game.current_night, game.current_day, and
current_day.current_vote). This adds some dummy types in
lib/typings/context.py which require each of those attributes to take
their respective type. They should be propagated through the relevant call
stack (and casted when relevant checks/modifications are performed) to
fix this issue. The remaining outstanding issue with this implementation
is that mypy seems to allow Context objects to be passed to functions
which require one of these new subtypes. I think the issue is that the
type is currently contravariant; it needs to be made invariant somehow.

I choose this approach over a couple other options:
- Setting mypy strict optional checking off -- I want strict optional
checking in other cases, including in the same files.
- Propagating is not none assertions throughout the codebase. This
seemed like it would add a lot of unnecessary complexity, given how many
such assertions there would need to be.
- Using TypeVars and Generic types -- I couldn't get this to work. I
tried making BOTCBot a generic type with a typevar constrained to
Optional[Game] and Game and mypy wouldn't allow the constructor to
assign self.game = None because it inferred the type as Game. I'm really
not sure why its inference overruled the explicit type annotation.