Require a value for Character.parent
This is a temporary fix to #4 which basically just flips the issue (so
that passing None to characters violates static type checking) and then
ignores the type checker when we need to do that. This is just so when I
make other edits on this branch, I don't need to navigate through all
the mypy errors relating to this issue in order to find the other linter
errors.